The security is implemented in the backend
1) The classes are responsible for security logic are in the package com.mlreef.rest.security
2) The main class that calculates permissions - com.mlreef.rest.security.MlReefSecurityExpressionRoot. It can be extended with new methods for any security checks
3) GroupsController and DataProjectController/CodeProjectsController are covered with the new Security
4) TokenDetails is extended with 3 properties - username, groups, projects that allow to calculate permissions on the fly
5) TokenDetails is being placed to the Session cache at first request. Spring's SessionCache is extended with new logic that allows to store TokenDetails in cache by username also
6) com.mlreef.rest.security.MlReefSessionRepository that extends Spring's RedisSessionRepository for additional logic (ses previous item)
7) GroupService is added that covers group management
8) GroupController is added for REST access to group management
9) Session service is added to control and test sessions in backend
10) Ansync Pub/Sub possibilities added to the project using Redis as MessageBus. The business logic for that as example can be found in com.mlreef.rest.security.annotations.RefreshUserInformation and GroupsService.
11) Tables are extended with new fields. Subject entities have a new gitlab_id column now.
12) !!! gitlab_id column from Account entity must not be used. Replacement is Person.gitlabId. The column and field will be removed soon

The new rules that should be discussed and followed if ok:

1) Not any permission checks on the services layer. All methods of the services should be called without any owner information (token, ownerId etc.). Permissions check should be on API layer only
2) No any repositories placed in API/REST layer. Methods calls should go through service layer.
3) All Gitlab operations must be called under Admin account (all user's permissions who called Gitlab api must be checked before).Exceptions: group, project creation, because when you create a group under user account gitlab assign owner of the group to the calling user.

P.S. All SQL scripts will be renamed with version change after merge with development branch

+++ mlreef-rest/src/main/kotlin/com/mlreef/rest/feature/pipeline/PipelineService.kt
@@ -26,7 +26,7 @@ import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
-import java.util.*
+import java.util.UUID
 import java.util.UUID.randomUUID
 
 @Service
@@ -92,7 +92,7 @@ class PipelineService(
         return YamlFileGenerator().generateYamlFile(author, dataProject, gitlabRootUrl, pipelineInstance.sourceBranch, pipelineInstance.targetBranch, pipelineInstance.dataOperations)
     }
 
-    fun commitYamlFile(userToken: String, projectId: Int, targetBranch: String, fileContent: String, sourceBranch: String = "master"): Commit {
+    fun commitYamlFile(userToken: String, projectId: Long, targetBranch: String, fileContent: String, sourceBranch: String = "master"): Commit {
         val commitMessage = "pipeline execution"
         val fileContents = mapOf(Pair(".mlreef.yml", fileContent))
         try {
@@ -113,7 +113,7 @@ class PipelineService(
         }
     }
 
-    fun deletePipelineInstance(userToken: String, gitlabProjectId: Int, targetBranch: String) {
+    fun deletePipelineInstance(userToken: String, gitlabProjectId: Long, targetBranch: String) {
         try {
             gitlabRestClient.deleteBranch(token = userToken, projectId = gitlabProjectId, targetBranch = targetBranch)
         } catch (e: RestException) {
@@ -122,7 +122,7 @@ class PipelineService(
         log.info("PipelineInstance deleted: $targetBranch")
     }
 
-    fun startInstance(author: Account, userToken: String, gitlabProjectId: Int, instance: PipelineInstance): PipelineInstance {
+    fun startInstance(author: Account, userToken: String, gitlabProjectId: Long, instance: PipelineInstance): PipelineInstance {
         val fileContent = createPipelineInstanceFile(author = author, pipelineInstance = instance)
 
         commitYamlFile(
